base::make_span requires explicit size casts
Chromium change:
https://chromium.googlesource.com/chromium/src/+/e61641f724481fba9c20cb09388c537ff6c203a5

commit e61641f724481fba9c20cb09388c537ff6c203a5
Author: Will Harris <wfh@chromium.org>
Date:   Tue Feb 7 19:35:02 2023 +0000

    Enable explicit conversion checks for base::span sizes.

    This CL adds compile-time checks that the size parameter passed
    to base::span either by the constructor or the base::make_span
    function will fit into a size_t type and are not unsigned.

    This is achieved by use of base::StrictNumeric parameter to the
    methods and constructor.

    This has no runtime overhead, as the checks are done at compile
    time.

    This CL also contains all the remaining mechanical changes needed
    to convert signed numeric literal values to unsigned by adding
    a 'u' on the end. Any functional or more complex changes have
    already landed in other CLs tagged to this bug with separate
    owner review.

    This CL also adds no-compile tests for the new APIs.

    BUG=1385166

+++ ios/browser/api/net/certificate_utility.mm
@@ -73,7 +73,8 @@ namespace {
 
   bssl::UniquePtr<CRYPTO_BUFFER> cert_buffer(
       net::X509Certificate::CreateCertBufferFromBytes(base::make_span(
-          CFDataGetBytePtr(cert_data), CFDataGetLength(cert_data))));
+          CFDataGetBytePtr(cert_data),
+          base::checked_cast<size_t>(CFDataGetLength(cert_data)))));
 
   if (!cert_buffer) {
     return nil;
@@ -97,7 +98,8 @@ namespace {
 
   bssl::UniquePtr<CRYPTO_BUFFER> cert_buffer(
       net::X509Certificate::CreateCertBufferFromBytes(base::make_span(
-          CFDataGetBytePtr(cert_data), CFDataGetLength(cert_data))));
+          CFDataGetBytePtr(cert_data),
+          base::checked_cast<size_t>(CFDataGetLength(cert_data)))));
 
   if (!cert_buffer) {
     return nil;
